# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /integrations/slack/commands:
    post:
      tags: [Integrations]
      summary: Slack slash command
      operationId: slackCommand
      security: []
      description: |
        Endpoint for a Slack slash command such as `/sorolens <contract_id>`.
        Every request must carry a valid Slack signature
        (`X-Slack-Signature`, HMAC-SHA256 over `v0:<timestamp>:<body>` with the
        app's signing secret) and an `X-Slack-Request-Timestamp` within five
        minutes. Replies with an ephemeral Block Kit message showing the
        contract's watchdog status and latest alerts. Disabled (404) unless
        `SLACK_SIGNING_SECRET` is set.
      parameters:
        - name: X-Slack-Signature
          in: header
          required: true
          schema:
            type: string
            example: v0=a2114d57b48eac39b9ad189dd8316235a7b4a8d21a10bd27519666489c69b503
        - name: X-Slack-Request-Timestamp
          in: header
          required: true
          schema:
            type: string
            example: "1531420618"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                command:
                  type: string
                  example: /sorolens
                text:
                  type: string
                  description: Contract ID to look up; empty or `help` returns usage.
      responses:
        "200":
          description: Ephemeral Slack message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlackMessage"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          description: Missing, invalid, or stale Slack signature.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Slack integration is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          headers:
            X-Cache:
              $ref: "#/components/headers/XCache"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/watchdog/contracts/{id}:
    get:
      tags: [Watchdog]
      summary: Get one monitored contract
      operationId: getMonitoredContract
      security:
        - ApiKeyAuth: [read:watchdog]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The monitored contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoredContract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
  parameters:
    ContractID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Contract ID (starts with C).
...
  headers:
    XCache:
      description: |
        Response cache status (`HIT` or `MISS`). Present when the API runs
        with Redis caching enabled; cached bodies live for API_CACHE_TTL
        (default 30s) and are purged by writes to the same resource.
      schema:
        type: string
        enum: [HIT, MISS]
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidInput:
      description: Invalid query or body input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: object
              required: [code, message]
              properties:
                code:
                  type: string
                  enum: [NOT_FOUND, INVALID_INPUT, INTERNAL, RATE_LIMITED, UNSUPPORTED_MEDIA_TYPE]
                message:
                  type: string
                request_id:
                  type: string
            - type: object
              properties:
                error:
                  type: string
...
    Contract:
      type: object
      required: [id, network, label, wasm_hash, created_at_ledger, backfill_complete_at, status, added_at]
      properties:
        id:
          type: string
        network:
          type: string
        label:
          type: string
        wasm_hash:
          type: string
        created_at_ledger:
          type: integer
          format: int64
        backfill_complete_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, active, backfilling, error, paused]
        added_at:
          type: string
          format: date-time
...
    MonitoredContract:
      type: object
      required: [contract_id, network, name, owner, status, last_check, check_interval, registered_at, updated_at]
      properties:
        contract_id:
          type: string
        network:
          type: string
        name:
          type: string
        owner:
          type: string
        status:
          type: string
        last_check:
          type: string
          format: date-time
          nullable: true
        check_interval:
          type: integer
          format: int64
        registered_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    SlackMessage:
      type: object
      required: [response_type, text]
      properties:
        response_type:
          type: string
          enum: [ephemeral]
        text:
          type: string
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true